# Java/swagger.yml
swagger: "2.0"
info:
  title: Swagger Hire-it
  version: '1.0'
...
paths:
...
  /export/demanda:
    post:
      tags:
      - Exportacao
      summary: "Exporta demandas"
      description: "Faz uma exportação das demandas do usuário para uma lista"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/ExportacaoDemanda"
      responses:
        '200':
          description: Ok 
        '400':
          description: Erro ao criar a demanda    
  /export/oferta:
    post:
      tags:
      - Exportacao
      summary: "Exporta ofertas"
      description: "Faz uma exportação das ofertas do usuário para uma lista"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/ExportacaoOferta"
      responses:
        '200':
          description: Ok 
        '400':
          description: Erro ao criar a oferta     
...
  /localizacoes:
    post:
      tags:
      - Localizações
      summary: "Cadastra uma localização"
      description: "Cadastra uma localização"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/Localizacoes"
      responses:
        '201':
          description: Created
        '422':
          description: Informações inválidas
    get:
      tags:
      - Localizações
      summary: "Retorna todas localizacoes"
      description: "Retorna todas localizacoes"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/LocalizacoesResponse"
        '204':
          description: No content
  /localizacoes/cep/{cep}:
    get:
      tags:
      - Localizações
      summary: "Busca uma localização com base em um CEP"
      description: "Busca uma localização com base em um CEP"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/LocalizacoesCepResponse"
        '204':
          description: No content
  /localizacoes/{idLocalizacao}:
    delete:
      tags:
        - Localizações
      summary: "Deleta uma localização com base em um ID"
      description: "Deleta uma localização com base em um ID"
      parameters:
        - name: "idLocalizacao"
          in: "path"
          description: "ID da localizacao que deve ser deletada"
          required: true
          type: "integer"
          format: "int64"
      responses:
        '204':
          description: No content
        '404':
          description: Not Found
    put:
      tags:
        - Localizações
      summary: "Atualiza uma localização com base em um ID"
      description: "Atualiza uma localização com base em um ID"
      parameters:
        - name: "idLocalizacao"
          in: "path"
          description: "ID da localizacao que deve ser atualizada"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "body"
          schema:
            $ref: "#/definitions/Localizacoes"
      responses:
        '204':
          description: No content
        '404':
          description: Not Found
        '422':
          description: Informações inválidas
  /ofertas:
    post:
      tags:
      - Oferta
      summary: "Cria uma oferta"
      description: "Cria uma oferta com tecnologias"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Oferta"
      responses:
        '202':
          description: Accepted
        '422':
          description: Informações inválidas
    get:
      tags:
      - Oferta
      summary: "Retorna todas ofertas"
      description: "Retorna todas ofertas"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/OfertaResponse"
        '204':
          description: No content
  /ofertas/{idOferta}:
    get:
      tags:
      - Oferta
      summary: "Retorna uma oferta"
      description: "Retorna uma oferta com base no ID"
      parameters:
        - name: "idOferta"
          in: "path"
          description: "ID da oferta que deve retornar"
          required: true
          type: "integer"
          format: "int64"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/OfertaIdResponse"
        '204':
          description: No content
  /ofertas/usuario/{idUsuario}:
    get:
      tags:
      - Oferta
      summary: "Retorna todas ofertas de um usuário"
      description: "Retorna todas ofertas de um usuário"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/OfertaUsuarioResponse"
        '204':
          description: No content
  /oferta/filtro:
    get:
        tags:
        - Oferta
        summary: "Retorna todas ofertas filtradas"
        description: "Retorna todas ofertas com base nas informações passadas no filtro"
        parameters:
          - in: "body"
            name: "body"
            description: "Todos os campos são opcionais"
            required: false
            schema:
              $ref: "#/definitions/DemandaFiltro"
        responses:
          '200':
            description: Ok
            schema:
              $ref: "#/definitions/OfertaFiltroResponse"
          '204':
            description: No content
...
  /tecnologias/oferta/{idOferta}:
    get:
      tags:
      - Tecnologia
      summary: "Retorna todas as tecnologias de uma oferta"
      description: "Retorna todas as tecnologias de uma oferta"
      responses:
        '200':
          description: Ok
          schema:
            $ref: "#/definitions/TecnologiaOfertaResponse"
        '204':
          description: No content
  /tecnologias/oferta:
    post:
      tags:
      - Tecnologia
      summary: "Adiciona uma tecnologia de uma oferta"
      description: "Adiciona uma tecnologia de uma oferta"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Tecnologia"
      responses:
        '201':
          description: Created
        '422':
          description: Informações inválidas
...
  /usuario/{idUsuario}:
    get:
        tags:
        - Usuário
        summary: "Retorna um usuário"
        description: "Retorna um usuário de acordo com o ID no Path"
        parameters:
        - name: "idUsuário"
          in: "path"
          description: "ID do usuário que deve retornar"
          required: true
          type: "integer"
          format: "int64"
        responses:
          '200':
            description: Ok
            schema:
              $ref: "#/definitions/UsuarioIdResponse"
          '204':
            description: No content
    delete:
      tags:
        - Usuário
      summary: "Deleta um usuário com base em um ID"
      description: "Deleta um usuário com base em um ID"
      parameters:
        - name: "idUsuario"
          in: "path"
          description: "ID do usuário que deve ser deletado"
          required: true
          type: "integer"
          format: "int64"
      responses:
        '204':
          description: No content
        '404':
          description: Not Found
    put:
      tags:
        - Usuário
      summary: "Atualiza um usuário com base em um ID"
      description: "Atualiza um usuário com base em um ID"
      parameters:
        - name: "idUsuario"
          in: "path"
          description: "ID do usuário que deve ser atualizado"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "body"
          schema:
            $ref: "#/definitions/Usuario"
      responses:
        '204':
          description: No content
        '404':
          description: Not Found
        '422':
          description: Informações inválidas
...
definitions:
...
    DemandaFiltro:
      type: "object"
      properties:
        titulo: 
          type: string
          example: Primeira
        data: 
          type: string
          example: 2021-06-05T16:27:57.36
        UF: 
          type: string
          example: SP
        salarioMin: 
          type: number
          example: 2400
        salarioMax: 
          type: number
          example: 5000
        usuario: 
          type: string
          example: Thiago
        tecnologia: 
          type: string
          example: Java
        experiencia: 
          type: number
          example: 1
...
    ExportacaoDemanda:
      type: "object"
      properties: 
        titulo: 
          type: string
          example: Minha demanda
        descricao: 
          type: string
          example: Essa é a demanda que criei
        salario: 
          type: number
          example: 1000
        usuario:
          properties: 
            idUsuario:
              type: number
              example: 1
    ExportacaoOferta:
      type: "object"
      properties: 
        descricao: 
          type: string
          example: Essa é a oferta que criei
        usuario:
          properties: 
            idUsuario:
              type: number
              example: 1
    LocalizacoesResponse:
      type: "array"
      example: [
        {
            "idLocalizacao": 1,
            "cidade": "São Caetano do Sul",
            "uf": "SP",
            "cep": "09572520"
        },
        {
            "idLocalizacao": 2,
            "cidade": "São Paulo",
            "uf": "SP",
            "cep": "00135001"
        }
      ]
    LocalizacoesCepResponse:
      type: "object"
      properties: 
        idLocalizacao: 
          type: number
          example: 1
        cidade: 
          type: string
          example: São Caetano do Sul
        uf: 
          type: string
          example: SP
        cep: 
          type: string
          example: 09572520
    Localizacoes:
      type: "object"
      properties:  
        cidade: 
          type: string
          example: São Caetano do Sul
        uf: 
          type: string
          example: SP
        cep: 
          type: string
          example: 09572520
    OfertaResponse:
      type: "array"
      example: [
        {
            "idOferta": 2,
            "descricao": "Essa é a primeira demanda",
            "createdAt": "2021-06-04T18:49:12.613",
            "usuario": {
                "idUsuario": 3,
                "nome": "Thiago Barbosa",
                "email": "thiaguinho@gmail.com",
                "senha": "senha123",
                "descricao": "DBA",
                "classificacao": 5.0,
                "telefone": "962345673",
                "localizacao": {
                    "idLocalizacao": 1,
                    "cidade": "São Caetano do Sul",
                    "uf": "SP",
                    "cep": "09572520"
                }
            }
        },
        {
            "idOferta": 3,
            "descricao": "Essa é a primeira demanda",
            "createdAt": "2021-06-04T18:50:54.98",
            "usuario": {
                "idUsuario": 3,
                "nome": "Thiago Barbosa",
                "email": "thiaguinho@gmail.com",
                "senha": "senha123",
                "descricao": "DBA",
                "classificacao": 5.0,
                "telefone": "962345673",
                "localizacao": {
                    "idLocalizacao": 1,
                    "cidade": "São Caetano do Sul",
                    "uf": "SP",
                    "cep": "09572520"
                }
            }
        }
      ]
    OfertaIdResponse:
      type: "object"
      properties:
        idOferta: 
          type: number
          example: 3
        descricao: 
          type: string
          example: Essa é a primeira demanda
        createdAt: 
          type: string
          example: 2021-06-04T18:50:54.98
        usuario: 
          type: object
          properties: 
            idUsuario: 
              type: number
              example: 3
            nome: 
              type: string
              example: Thiago Barbosa
            email: 
              type: string
              example: thiaguinho@gmail.com
            senha: 
              type: string
              example: senha123
            descricao: 
              type: string
              example: DBA
            classificacao: 
              type: number
              example: 5
            telefone: 
              type: string
              example: 962345673
            localizacao: 
              type: object
              properties: 
                idLocalizacao: 
                  type: number
                  example: 1
                cidade: 
                  type: string
                  example: São Caetano do Sul
                uf: 
                  type: string
                  example: SP
                cep: 
                  type: string
                  example: 09572520
    OfertaUsuarioResponse:
      type: array
      example: [
        {
            "idOferta": 2,
            "descricao": "Essa é a primeira demanda",
            "createdAt": "2021-06-04T18:49:12.613",
            "usuario": {
                "idUsuario": 3,
                "nome": "Thiago Barbosa",
                "email": "thiaguinho@gmail.com",
                "senha": "senha123",
                "descricao": "DBA",
                "classificacao": 5.0,
                "telefone": "962345673",
                "localizacao": {
                    "idLocalizacao": 1,
                    "cidade": "São Caetano do Sul",
                    "uf": "SP",
                    "cep": "09572520"
                }
            }
        },
        {
            "idOferta": 3,
            "descricao": "Essa é a primeira demanda",
            "createdAt": "2021-06-04T18:50:54.98",
            "usuario": {
                "idUsuario": 3,
                "nome": "Thiago Barbosa",
                "email": "thiaguinho@gmail.com",
                "senha": "senha123",
                "descricao": "DBA",
                "classificacao": 5.0,
                "telefone": "962345673",
                "localizacao": {
                    "idLocalizacao": 1,
                    "cidade": "São Caetano do Sul",
                    "uf": "SP",
                    "cep": "09572520"
                }
            }
        }
      ]
    Oferta:
      type: object
      properties:
        oferta: 
          type: object
          properties: 
            descricao: 
              type: string
              example: Essa é a primeira demanda
            usuario: 
              type: object
              properties: 
                idUsuario: 
                  type: number
                  example: 3
        listaTecnologias: 
          type: array
          items: 
            type: object
            properties: 
              tempoExperiencia: 
                type: number
                example: 10
              tecnologias: 
                type: object
                properties: 
                  idTecnologia: 
                    type: number
                    example: 1
    OfertaFiltroResponse:
      type: array
      example: [
          {
            "idOferta": 2,
            "descricao": "Essa é a primeira demanda",
            "createdAt": "2021-06-04T18:49:12.613",
            "usuario": {
                "idUsuario": 3,
                "nome": "Thiago Barbosa",
                "email": "thiaguinho@gmail.com",
                "senha": "senha123",
                "descricao": "DBA",
                "classificacao": 5.0,
                "telefone": "962345673",
                "localizacao": {
                    "idLocalizacao": 1,
                    "cidade": "São Caetano do Sul",
                    "uf": "SP",
                    "cep": "09572520"
                }
            }
        },
        {
            "idOferta": 3,
            "descricao": "Essa é a primeira demanda",
            "createdAt": "2021-06-04T18:50:54.98",
            "usuario": {
                "idUsuario": 3,
                "nome": "Thiago Barbosa",
                "email": "thiaguinho@gmail.com",
                "senha": "senha123",
                "descricao": "DBA",
                "classificacao": 5.0,
                "telefone": "962345673",
                "localizacao": {
                    "idLocalizacao": 1,
                    "cidade": "São Caetano do Sul",
                    "uf": "SP",
                    "cep": "09572520"
                }
            }
        }
      ]
...
    TecnologiaOfertaResponse:
      type: "array"
      example: [
        {
            "idTecnologiaOferta": 1,
            "tempoExperiencia": 10,
            "tecnologias": {
                "idTecnologia": 1,
                "tecnologia": "Java"
            },
            "ofertas": {
                "idOferta": 2,
                "descricao": "Essa é a primeira demanda",
                "createdAt": "2021-06-04T18:49:12.613",
                "usuario": {
                    "idUsuario": 3,
                    "nome": "Thiago Barbosa",
                    "email": "thiaguinho@gmail.com",
                    "senha": "senha123",
                    "descricao": "DBA",
                    "classificacao": 5.0,
                    "telefone": "962345673",
                    "localizacao": {
                        "idLocalizacao": 1,
                        "cidade": "São Caetano do Sul",
                        "uf": "SP",
                        "cep": "09572520"
                    }
                }
            }
        },
        {
            "idTecnologiaOferta": 3,
            "tempoExperiencia": 10,
            "tecnologias": {
                "idTecnologia": 2,
                "tecnologia": "Springboot"
            },
            "ofertas": {
                "idOferta": 2,
                "descricao": "Essa é a primeira demanda",
                "createdAt": "2021-06-04T18:49:12.613",
                "usuario": {
                    "idUsuario": 3,
                    "nome": "Thiago Barbosa",
                    "email": "thiaguinho@gmail.com",
                    "senha": "senha123",
                    "descricao": "DBA",
                    "classificacao": 5.0,
                    "telefone": "962345673",
                    "localizacao": {
                        "idLocalizacao": 1,
                        "cidade": "São Caetano do Sul",
                        "uf": "SP",
                        "cep": "09572520"
                    }
                }
            }
        }
      ]
    Tecnologia:
      type: "object"
      properties:
        tempoExperiencia: 
          type: number
          example: 10
        tecnologias: 
          type: object
          properties: 
            idTecnologia: 
              type: number
              example: 1
        ofertas: 
          type: object
          properties: 
            idOferta: 
              type: number
              example: 4
    Usuario:
      type: "object"
      properties: 
        nome: 
          type: string
          example: Mateus Gomes
        email: 
          type: string
          example: mateusgomes@gmail.com
        senha: 
          type: string
          example: senha123
        descricao: 
          type: string
          example: Backend Developer
        telefone: 
          type: string
          example: 962345673
        localizacao: 
          type: object
          properties: 
            idLocalizacao: 
              type: number
              example: 2
...
    UsuarioIdResponse:
      type: object
      properties:
        idUsuario: 
          type: number
          example: 3
        nome: 
          type: string
          example: Thiago Barbosa
        email: 
          type: string
          example: thiaguinho@gmail.com
        senha: 
          type: string
          example: senha123
        descricao: 
          type: string
          example: DBA
        classificacao: 
          type: number
          example: 5
        telefone: 
          type: string
          example: 962345673
        localizacao: 
          type: object
          properties: 
            idLocalizacao: 
              type: number
              example: 1
            cidade: 
              type: string
              example: São Caetano do Sul
            uf: 
              type: string
              example: SP
            cep: 
              type: string
              example: 09572520
...